First implementation of extended configurability [major]
NOTE This disables the `init` and `update` sub-commands, for now.

This allows to configure which checks are run and in which order.
It also contains part of what is needed to configure settings of a check,
if it had any.
None of them has any yet though,
and I would prefer it to stay like this as much as possible,
as the goal is to have unilateral standards,
not a myriad of different settings followed by different projects.